requirements.txt
asgiref==3.7.2
cfgv==3.3.1
distlib==0.3.7
Django==4.2
django-filter==23.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
identify==2.5.26
nodeenv==1.8.0
phonenumbers==8.13.18
Pillow==10.0.0
platformdirs==3.10.0
pre-commit==3.3.3
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
ruff==0.0.284
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.24.2
